Fixed MTLTexture generation on macOS

+++ Foil/Shaders.metal
@@ -41,14 +41,14 @@ vertex VertexOut image_vertex(
 	return outVertex;
 }
 
-fragment float4 image_fragment(
+fragment half4 image_fragment(
 							   VertexOut vertexIn [[ stage_in ]],
 							   constant Uniforms & uniforms [[ buffer(0) ]],
 							   constant float4* color [[ buffer(2) ]],
 							   texture2d<float, access::sample> colorTexture [[ texture(0) ]],
 							   sampler colorSampler [[ sampler(0) ]]
 							   ) {
-	return  float4(colorTexture.sample(colorSampler, vertexIn.texcoords).rgba) * color[0];
+	return half4(colorTexture.sample(colorSampler, vertexIn.texcoords).rgba * color[0]);
 }
